Refactor: Move all infra code to subdir
This way the main tuf_conformance/ directory contains just tests.
