+++ tuf_conformance/__init__.py
@@ -3,4 +3,4 @@ import pytest
 __version__ = "2.2.0"
 
 # register pytest asserts before the imports happen in conftest.py
-pytest.register_assert_rewrite("tuf_conformance.client_runner")
+pytest.register_assert_rewrite("tuf_conformance._internal.client_runner")
