+++ tuf_conformance/test_file_download.py
@@ -1,9 +1,9 @@
 import pytest
 from tuf.api.metadata import Snapshot, TargetFile, Targets, Timestamp
 
-from tuf_conformance.client_runner import ClientRunner
-from tuf_conformance.repository_simulator import Artifact
-from tuf_conformance.simulator_server import SimulatorServer
+from tuf_conformance._internal.client_runner import ClientRunner
+from tuf_conformance._internal.repository_simulator import Artifact
+from tuf_conformance._internal.simulator_server import SimulatorServer
 
 
 def test_client_downloads_expected_file(
