+++ tuf_conformance/test_keys.py
@@ -3,8 +3,8 @@ import json
 import pytest
 from tuf.api.metadata import Metadata, Root, Snapshot, Targets, Timestamp
 
-from tuf_conformance.client_runner import ClientRunner
-from tuf_conformance.simulator_server import SimulatorServer
+from tuf_conformance._internal.client_runner import ClientRunner
+from tuf_conformance._internal.simulator_server import SimulatorServer
 
 
 def test_snapshot_does_not_meet_threshold(
