+++ tuf_conformance/test_updater_key_rotations.py
@@ -3,8 +3,8 @@ from dataclasses import dataclass
 import pytest
 from tuf.api.metadata import Root, Snapshot, Targets, Timestamp
 
-from tuf_conformance.client_runner import ClientRunner
-from tuf_conformance.simulator_server import SimulatorServer
+from tuf_conformance._internal.client_runner import ClientRunner
+from tuf_conformance._internal.simulator_server import SimulatorServer
 
 
 @dataclass
